+++ src/SMAPI.Common/Platform.cs
@@ -1,0 +1,15 @@
+namespace StardewModdingAPI.Common
+{
+    /// <summary>The game's platform version.</summary>
+    internal enum Platform
+    {
+        /// <summary>The Linux version of the game.</summary>
+        Linux,
+
+        /// <summary>The Mac version of the game.</summary>
+        Mac,
+
+        /// <summary>The Windows version of the game.</summary>
+        Windows
+    }
+}