+++ src/SMAPI/Framework/ModLoading/AssemblyLoader.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
+using StardewModdingAPI.Common;
 using StardewModdingAPI.Framework.Exceptions;
 using StardewModdingAPI.Metadata;