+++ src/SMAPI/Framework/ModLoading/PlatformAssemblyMap.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Mono.Cecil;
+using StardewModdingAPI.Common;
 
 namespace StardewModdingAPI.Framework.ModLoading
 {