+++ docs/release-notes.md
@@ -14,6 +14,7 @@
   * Fixed issue where assets didn't reload correctly when the player switches language.
 
 * For SMAPI developers:
+  * Added MacOS detection to `Constants.Platform`.
   * Added prerelease versions to the mod update-check API response where available (GitHub only).
   * Added support for beta releases on the home page.
   * Split mod DB out of `StardewModdingAPI.config.json`, so we can load config earlier and reduce unnecessary memory usage later.